feat: Introduce IWorkItemQueryWindowStrategy and WorkItemQueryWindow classes for date-window WIQL chunking
