+++ src/DevOpsMigrationPlatform.Infrastructure.AzureDevOps/Services/CatalogService.cs
@@ -1,24 +1,24 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using DevOpsMigrationPlatform.Abstractions;
 using DevOpsMigrationPlatform.Abstractions.Services;
-using Microsoft.TeamFoundation.Core.WebApi;
-using Microsoft.VisualStudio.Services.Common;
-using Microsoft.VisualStudio.Services.WebApi;
 
 namespace DevOpsMigrationPlatform.Infrastructure.AzureDevOps.Services;
 
 public class CatalogService : ICatalogService
 {
     private readonly IWorkItemDiscoveryService _workItemDiscovery;
+    private readonly IProjectDiscoveryService _projectDiscovery;
 
-    public CatalogService(IWorkItemDiscoveryService workItemDiscovery)
+    public CatalogService(
+        IWorkItemDiscoveryService workItemDiscovery,
+        IProjectDiscoveryService projectDiscovery)
     {
         _workItemDiscovery = workItemDiscovery ?? throw new ArgumentNullException(nameof(workItemDiscovery));
+        _projectDiscovery = projectDiscovery ?? throw new ArgumentNullException(nameof(projectDiscovery));
     }
 
     public async Task<IReadOnlyList<string>> GetProjectsAsync(
@@ -26,11 +26,8 @@ public class CatalogService : ICatalogService
         string pat,
         CancellationToken cancellationToken = default)
     {
-        var credentials = new VssBasicCredential(string.Empty, pat);
-        var connection = new VssConnection(new Uri(orgUrl), credentials);
-        var projectClient = connection.GetClient<ProjectHttpClient>();
-        var projects = await projectClient.GetProjects();
-        return projects.Select(p => p.Name).ToList();
+        var projects = await _projectDiscovery.GetProjectsAsync(orgUrl, pat, cancellationToken);
+        return projects;
     }
 
     public async IAsyncEnumerable<ProjectDiscoverySummary> CountAllWorkItemsAsync(
