+++ src/DevOpsMigrationPlatform.Infrastructure.AzureDevOps/Services/WorkItemQueryWindowStrategy.cs
@@ -4,33 +4,13 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
+using DevOpsMigrationPlatform.Abstractions.Services;
 using Microsoft.TeamFoundation.WorkItemTracking.WebApi;
 using Microsoft.TeamFoundation.WorkItemTracking.WebApi.Models;
 using Microsoft.VisualStudio.Services.Common;
 using Microsoft.VisualStudio.Services.WebApi;
 
 namespace DevOpsMigrationPlatform.Infrastructure.AzureDevOps.Services;
-
-/// <summary>
-/// Options controlling the date-window algorithm.
-/// </summary>
-public sealed class WorkItemQueryWindowOptions
-{
-    public int InitialWindowDays { get; set; } = 120;
-    public int LimitThreshold { get; set; } = 20_000;
-    public int MinWindowDays { get; set; } = 1;
-}
-
-/// <summary>
-/// One window result yielded by <see cref="WorkItemQueryWindowStrategy"/>.
-/// </summary>
-public sealed class WorkItemQueryWindow
-{
-    public DateTime WindowStart { get; init; }
-    public DateTime WindowEnd { get; init; }
-    public TimeSpan WindowSize { get; init; }
-    public IReadOnlyList<int> WorkItemIds { get; init; } = Array.Empty<int>();
-}
 
 /// <summary>
 /// Shared date-window WIQL strategy that keeps each query under the 20,000-item limit.
